Introduction

Tool for Quantitative Portfolio Analytics

Inspired by PerformanceAnalytics and PortoflioAnalytics packages in R, pyfolio in Python, PortfolioAnalytics.jl aims to provide users with functionality for performing portfolio analytics.

PortfolioAnalytics.jl is minimal now, but it is under heavy development.

The following functions are available:

Return()

PortfolioReturn()

SharpeRatio()

VaR()

PortfolioOptimize()

On the pipe:

MeanReturns()

Moments()

ExpectedShortfall()

Known issues

These are just a few examples of known issues with functions in the package. These issues will be fixed in the next release.

All Functions:

Types (e.g., TSFrame, Vector{Float64}) still need to be specified. This hasn't been done on purpose. The best candidate looks TSFrame but I'm still open to suggestions. Nevertheless, the types of parameters to be passed to the functions will be set in the next release.

PortfolioReturn()

It returns a vector instead of a proper object with dates. This issue is connected to #1 (All Functions) and will be fixed in the next release.

Sharpe Ratio()

Does not work in presense of NAs or missing vlaues. Will be fixed in the next release.

Needs to be passed single column only. if multiple columns are passed it will be calculaing one single SharpeRatio which is wrong. Will be fixed in the next release.

VaR()

It cannot run on multiple columns. Will be fixed in the next release.

Does not work in presense of NAs or missing vlaues. Will be fixed in the next release.

PortfolioOptimize()

Doesn't work in presense of NAs or missing values. Will be fixed in the next release.

TO DO

New functionalities will also be added to the existing functions.

Return()

***period*** parameter to be added for multi-period returns

PortfolioReturn()

***Rebalancing*** parameter will be added

***period*** parameter will be added for multi-period returns

VaR()

***VaR with Monte Carlo*** to be added

PortfolioOptimize()

***Maximization by SharpeRatio*** will be added

***Custom constraints*** functionality will be included.

***NamedArray*** to be added for weights.

Contributions are most welcome

We greatly value contributions of any kind. Contributions could include but are not limited to documentation improvements, bug reports, new or improved code, scientific and technical code reviews, infrastructure improvements, mailing lists, chat participation, community help/building, education, and outreach.

Bug reports

Please report any issues via the GitHub issue tracker. All kinds of issues are welcome and encouraged; this includes bug reports, documentation typos, feature requests, etc.

User Guide

Installation

julia> using Pkg
julia> Pkg.add("PortfolioAnalytics")

PortfolioAnalytics.MeanReturn — Method

Calculates the mean return from asset return

MeanReturn(returns)

Arguments: R: column(s) of TSFrame object of asset returns

Output: - NamedArray Notes: -

Issues: - Doesn't work in presense of NAs or missing values. Will be fixed in the next release

To do: -

PortfolioAnalytics.Moments — Method

Calculates the (statistical) moments of asset returns

Moments(R)

Arguments: R: column(s) of TSFrame object of asset returns

Output: - NamedArray, rows: moments, columns: tickers Notes: - Kurtosis: excess kurtosis

Issues: - Doesn't work in presense of NAs or missing values. Will be fixed in the next release

To do: -

PortfolioAnalytics.PortfolioOptimize — Method

Calculates the optimal Portfolio weights for minumum-variance portfolio

PortfolioOptimize(returns[2:13])

Arguments: R: columns of TSFrame object of asset returns

Output: - Named Tuple - 1, preturn : portfolio mean return - 2, pvar: portfolio variable - 3, pweights: optimal portfolio weights for minumum variance portfolio Notes: -

Issues: - Doesn't work in presense of NAs or missing values. Will be fixed in the next release

To do: - Maximize SharpeRatio - Custom constraints - NamedArray to be used for weights

PortfolioAnalytics.PortfolioReturn — Function

Calculates portfolio return

preturns = PortfolioReturn(prices_ts, [0.4, 0.4, 0.2])

Arguments: - price: column(s) of TSFrame object of asset prices - weights: weights of assets of type Vector{Float64} - period: Period, Int

Output: - TSFrame object, missing period is automatically removed

Notes: -

Issues: -

To do: - Rebalancing to be added

PortfolioAnalytics.Return — Function

Return(price::TSFrame, period::Int=1)

Calculates returns form asset prices

Arguments: - price: column(s) of TSFrame object of asset prices - period: period, Int

Output: - TSFrame object, missing period is automatically removed

Notes: - This function does not calculate the portfolio return. Portfolio return is calcuated by PortfolioReturn function.

Issues: -

To do: -

PortfolioAnalytics.SharpeRatio — Function

Calculates Sharpe Ratio

preturns_ts = TSFrame(preturns[2:end])
SharpeRatio(preturns_ts)

Arguments: - R: column(s) of TSFrame object of asset returns - Rf: Risk free rate, default = 0

Output: - NamedArray

Notes: -

Issues: - It does not accept the missing values. Will be fixed in the next release. - Needs to be passed single column only. if multiple columns are passed it will be calculaing one single SharpeRatio which is wrong. Will be fixed in the next release.

To do: -

PortfolioAnalytics.VaR — Function

Calculates Value-at-Risk(VaR)

Arguments: - R: column(s) of TSFrame object - p: confidence interval, Number - method: method of VaR calculation, default = "historical", available methods: "historical" and "parametric", String

Output: - NamedArray

Notes: -

Issues: - Does not work in presense of NAs or missing vlaues. Will be fixed in the next release.

To do: - VaR with Monte Carlo to be implemented